+++ noodles-vcf/src/variant/record_buf/builder.rs
@@ -5,7 +5,7 @@ use noodles_core::Position;
 use super::{AlternateBases, Filters, Ids, Info, RecordBuf, Samples};
 
 /// A VCF record builder.
-#[derive(Debug, Default, PartialEq)]
+#[derive(Debug, PartialEq)]
 pub struct Builder {
     reference_sequence_name: String,
     variant_start: Option<Position>,
@@ -244,6 +244,22 @@ impl Builder {
     }
 }
 
+impl Default for Builder {
+    fn default() -> Self {
+        Self {
+            reference_sequence_name: String::from("."),
+            variant_start: Some(Position::MIN),
+            ids: Ids::default(),
+            reference_bases: String::from("N"),
+            alternate_bases: AlternateBases::default(),
+            quality_score: None,
+            filters: Filters::default(),
+            info: Info::default(),
+            samples: Samples::default(),
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -253,10 +269,10 @@ mod tests {
     fn test_default() {
         let record = Builder::default();
 
-        assert!(record.reference_sequence_name.is_empty());
-        assert!(record.variant_start.is_none());
+        assert_eq!(record.reference_sequence_name, ".");
+        assert_eq!(record.variant_start, Some(Position::MIN));
         assert!(record.ids.as_ref().is_empty());
-        assert!(record.reference_bases.is_empty());
+        assert_eq!(record.reference_bases, "N");
         assert!(record.alternate_bases.is_empty());
         assert!(record.quality_score.is_none());
         assert!(record.filters.as_ref().is_empty());