+++ noodles-vcf/CHANGELOG.md
@@ -9,6 +9,10 @@
     This previously created a new parser, which may not have the same applied
     options.
 
+  * vcf/variant/record_buf/builder: Initialize default builder.
+
+    The default builder now matches a default `RecordBuf`.
+
 ## 0.89.0 - 2026-07-10
 
 ### Changed